refactor: move app folder to lib
