+++ src/components/home-hero-section.tsx
@@ -1,9 +1,9 @@
 import Link from 'next/link'
 import { Fragment } from 'react'
 
-import { useI18n } from '@/app/i18n/use-i18n'
-import { useAppMetadata } from '@/app/metadata/use-app-metadata'
-import * as routes from '@/app/route/routes'
+import { useI18n } from '@/lib/i18n/use-i18n'
+import { useAppMetadata } from '@/lib/metadata/use-app-metadata'
+import * as routes from '@/lib/route/routes'
 
 export function HomeHeroSection(): JSX.Element {
   const { t } = useI18n<'common'>()
