+++ src/lib/analytics/analytics-script.tsx
