+++ src/lib/i18n/common/index.ts
@@ -1,6 +1,6 @@
-import type { Plurals as _Plurals, Plurals } from '@/app/i18n/dictionaries'
 import { type Gender } from '@/db/genders'
 import { type EntityKind } from '@/db/types'
+import type { Plurals as _Plurals, Plurals } from '@/lib/i18n/dictionaries'
 import type { Locale } from '~/config/i18n.config'
 
 export interface Dictionary {
