+++ src/lib/i18n/use-i18n.ts
@@ -1,7 +1,7 @@
 import type { I18nService } from '@stefanprobst/next-i18n'
 import { useI18n as useInternationalisation } from '@stefanprobst/next-i18n'
 
-import type { Dictionaries } from '@/app/i18n/dictionaries'
+import type { Dictionaries } from '@/lib/i18n/dictionaries'
 import type { Locale } from '~/config/i18n.config'
 
 export function useI18n<K extends keyof Dictionaries = never>(): I18nService<
