+++ src/lib/route/use-canonical-url.ts
@@ -1,9 +1,9 @@
 import type { UrlSearchParamsInit } from '@stefanprobst/request'
 import { useMemo } from 'react'
 
-import { useLocale } from '@/app/route/use-locale'
-import { usePathname } from '@/app/route/use-pathname'
 import { createAppUrl } from '@/lib/create-app-url'
+import { useLocale } from '@/lib/route/use-locale'
+import { usePathname } from '@/lib/route/use-pathname'
 
 export function useCanonicalUrl(searchParams?: UrlSearchParamsInit): URL {
   const { pathname } = usePathname()
