+++ src/pages/_app.page.tsx
@@ -9,19 +9,19 @@ import type { NextWebVitalsMetric } from 'next/app'
 import Head from 'next/head'
 import { Fragment } from 'react'
 
-import { AnalyticsScript } from '@/app/analytics/analytics-script'
-import { reportPageView } from '@/app/analytics/analytics-service'
-import type { AppProps, GetLayout } from '@/app/app.types'
-import { useAppMetadata } from '@/app/metadata/use-app-metadata'
-import { PageLayout } from '@/app/page.layout'
-import { Providers } from '@/app/providers.context'
-import { RootErrorBoundaryFallback } from '@/app/root-error-boundary-fallback'
-import { useAlternateLanguageUrls } from '@/app/route/use-alternate-language-urls'
-import { useCanonicalUrl } from '@/app/route/use-canonical-url'
-import { useLocale } from '@/app/route/use-locale'
-import { usePageLoadProgressIndicator } from '@/app/use-page-load-progress-indicator'
+import { PageLayout } from '@/components/page.layout'
+import { RootErrorBoundaryFallback } from '@/components/root-error-boundary-fallback'
+import { AnalyticsScript } from '@/lib/analytics/analytics-script'
+import { reportPageView } from '@/lib/analytics/analytics-service'
+import type { AppProps, GetLayout } from '@/lib/app.types'
 import { createAppUrl } from '@/lib/create-app-url'
 import { createAssetLink } from '@/lib/create-asset-link'
+import { useAppMetadata } from '@/lib/metadata/use-app-metadata'
+import { Providers } from '@/lib/providers.context'
+import { useAlternateLanguageUrls } from '@/lib/route/use-alternate-language-urls'
+import { useCanonicalUrl } from '@/lib/route/use-canonical-url'
+import { useLocale } from '@/lib/route/use-locale'
+import { usePageLoadProgressIndicator } from '@/lib/use-page-load-progress-indicator'
 import { manifestFileName, openGraphImageName } from '~/config/metadata.config'
 
 export default function App(props: AppProps): JSX.Element {
