+++ src/pages/index.page.tsx
@@ -1,11 +1,11 @@
 import { PageMetadata } from '@stefanprobst/next-page-metadata'
 import { Fragment } from 'react'
 
-import { useI18n } from '@/app/i18n/use-i18n'
-import { withDictionaries } from '@/app/i18n/with-dictionaries'
-import { usePageTitleTemplate } from '@/app/metadata/use-page-title-template'
 import { HomeHeroSection } from '@/components/home-hero-section'
 import { MainContent } from '@/components/main-content'
+import { useI18n } from '@/lib/i18n/use-i18n'
+import { withDictionaries } from '@/lib/i18n/with-dictionaries'
+import { usePageTitleTemplate } from '@/lib/metadata/use-page-title-template'
 
 export const getStaticProps = withDictionaries(['common'])
